+++ tests/parser.rs
@@ -2,8 +2,7 @@ extern crate ddnnf_lib;
 
 use ddnnf_lib::ddnnf::{node::NodeType::*, Ddnnf};
 use ddnnf_lib::parser;
-
-use rug::Integer;
+use num::BigInt;
 
 #[test]
 fn ddnnf_parsing_test() {
@@ -12,18 +11,18 @@ fn ddnnf_parsing_test() {
     let mut ddnnf_c2d: Ddnnf = parser::build_ddnnf("./tests/data/small_ex_c2d.nnf", None);
 
     assert_eq!(ddnnf_c2d.number_of_variables, 4);
-    assert_eq!(ddnnf_c2d.rc(), 4);
+    assert_eq!(ddnnf_c2d.rc(), BigInt::from(4));
     assert_eq!(ddnnf_c2d.nodes.len(), 12);
 
     assert_eq!(ddnnf_d4.number_of_variables, 4);
-    assert_eq!(ddnnf_d4.rc(), 4);
+    assert_eq!(ddnnf_d4.rc(), BigInt::from(4));
     assert_eq!(ddnnf_d4.nodes.len(), 18);
 
     let and_node = ddnnf_c2d.nodes.pop().unwrap();
     match and_node.ntype {
         And { children } => {
             assert_eq!(children.len(), 3_usize);
-            assert_eq!(and_node.count, Integer::from(4_u32))
+            assert_eq!(and_node.count, BigInt::from(4))
         }
         _ => panic!("Node isn't an and node"),
     }
@@ -32,7 +31,7 @@ fn ddnnf_parsing_test() {
     match or_node.ntype {
         Or { children } => {
             assert_eq!(children.len(), 2_usize);
-            assert_eq!(or_node.count, Integer::from(2_u32))
+            assert_eq!(or_node.count, BigInt::from(2))
         }
         _ => panic!("Node isn't an or node"),
     }
